TransportBackend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.7
django-environ==0.11.2
Pillow==10.4.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
